requirements.txt
altgraph==0.17
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
dj-database-url==0.5.0
dj-static==0.0.6
Flask==2.0.1
future==0.18.2
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
line-bot-sdk==1.19.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.0
pandas==1.2.5
pandas-datareader==0.9.0
pefile==2021.5.24
Pillow==8.2.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.25.1
scipy==1.7.0
seaborn==0.11.1
six==1.16.0
static3==0.7.0
urllib3==1.26.5
Werkzeug==2.0.1
